# docs/swagger.yaml
basePath: /
definitions:
...
  database.NFTStore:
    properties:
      address:
        type: string
      cid:
        type: string
      time:
        type: string
      tokenId:
        type: integer
      type:
        type: string
    type: object
  types.ListNFTRes:
    properties:
      length:
        type: integer
      nftInfos:
        items:
          $ref: '#/definitions/database.NFTStore'
        type: array
    type: object
...
  types.MintRes:
    properties:
      tokenID:
        type: integer
    type: object
...
  types.TweetNFTInfoRes:
    properties:
      images:
        items:
          type: string
        type: array
      link:
        type: string
      name:
        type: string
      postTime:
        type: integer
      tweet:
        type: string
    type: object
...
host: test-xs-api.memolabs.net
info:
  contact: {}
  description: This is the API document for xspace server.
  title: Xspace-Server API
  version: "1.0"
paths:
...
  /v1/nft/data/info:
    get:
      consumes:
      - application/json
      description: Get DataNFT content
      parameters:
      - description: Bearer YOUR_ACCESS_TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: DataNFT's id
        in: query
        name: tokenID
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: DataNFT binary content
          schema:
            type: file
        "400":
          description: Bad Request
          schema: {}
        "403":
          description: Forbidden
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      tags:
      - NFT
  /v1/nft/data/mint:
    post:
      consumes:
      - multipart/form-data
      description: Mint user's data into NFTs
      parameters:
      - description: Bearer YOUR_ACCESS_TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: User's data
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.MintRes'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      tags:
      - NFT
  /v1/nft/list:
    get:
      consumes:
      - application/json
      description: List all NFT information belonging to the user
      parameters:
      - description: Bearer YOUR_ACCESS_TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: Pages
        in: query
        name: page
        required: true
        type: string
      - description: The amount of data displayed on each page
        in: query
        name: size
        required: true
        type: string
      - description: NFT type (tweet for tweetNFT, data for dataNFT, tweetNFT and
          dataNFT will be all listed by default)
        in: query
        name: type
        type: string
      - description: Order rules (date_asc for sorting by creation time from smallest
          to largest, date_desc for sorting by creation time from largest to smallest,
          dealut is date_desc)
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ListNFTRes'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      tags:
      - NFT
  /v1/nft/tweet/info:
    get:
      consumes:
      - application/json
      description: Get Post TweetNFT content
      parameters:
      - description: Bearer YOUR_ACCESS_TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: TweetNFT's id
        in: query
        name: tokenID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.TweetNFTInfoRes'
        "400":
          description: Bad Request
          schema: {}
        "403":
          description: Forbidden
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      tags:
      - NFT
  /v1/nft/tweet/mint:
    post:
      consumes:
      - application/json
      description: Mint user's tweets into NFTs
      parameters:
      - description: Bearer YOUR_ACCESS_TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: User's twtter/x name
        in: body
        name: name
        required: true
        schema:
          type: string
      - description: The timestamp when the user posted the tweet
        in: body
        name: postTime
        required: true
        schema:
          type: integer
      - description: The text of the tweet(including emoji)
        in: body
        name: tweet
        required: true
        schema:
          type: string
      - description: The image url of the tweet
        in: body
        name: images
        required: true
        schema:
          type: string
      - description: The link to the tweet
        in: body
        name: link
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.MintRes'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      tags:
      - NFT
  /v1/nft/tweet/reply/info:
    get:
      consumes:
      - application/json
      description: Get Reply Tweet content
      parameters:
      - description: Bearer YOUR_ACCESS_TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: Tweet Content's cid
        in: query
        name: cid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.TweetNFTInfoRes'
        "400":
          description: Bad Request
          schema: {}
        "403":
          description: Forbidden
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      tags:
      - NFT
...
swagger: "2.0"